lib-gralloc-qsd8k: Add thread to enable async posting of display buffers
Change fb_post() to support async posting of buffers to FB/MDP drivers.
This allows post to return more quickly and also support swapInterval 0.
A thread is used to run sync with FB/MDP driver but async to callers of
fb_post() such as surfaceflinger. Also add support for swapInterval 0.